Fix all "cargo clippy" warnings

+++ core/src/mem.rs
@@ -9,7 +9,7 @@ pub trait Mem {
     fn mem_read_u16(&mut self, addr: u16) -> u16 {
         let lo = self.mem_read(addr) as u16;
         let hi = self.mem_read(addr.wrapping_add(1)) as u16;
-        (hi << 8) | (lo as u16)
+        (hi << 8) | lo
     }
 
     fn mem_write_u16(&mut self, addr: u16, value: u16) {
